Fix leaked cgroup hierarchies when jobacct_gather/cgroup enabled.
jobacct_gather/cgroup cleanup happened before SIGKILL was sent by proctrack
or switch plugins. The proctrack_g_signal() only signals those tasks for
which the parent is the step, but not grandchildren except when the
signal is a SIGKILL. Moving the cleanup after the SIGKILL ensures all
processes are gone when the hierarchy removal is done.

Bug 9429

Conflicts:
	NEWS