Fix rejecting jobs under MaxMemPerCPU when allocating more cpus than tasks
MaxMemPerCPU can cause a job to be auto adjusted increasing pn_min_cpus due
to memory limits. In that situation if we also request --ntasks-per-node the
job may be rejected, because we may end up with too many cpus allocated to
the job. There shouldn't be any problem in allocating more cpus and not
using them if we require it due to MaxMemPerCPU. This works with --exclusive
flag.

This patch detects if there are enough cpus depending on pn_min_cpus and
for each node then it picks the maximum between this number and the
required cpus depending upon the job request.

Bug 9716

Conflicts:
	NEWS

REMOVE FOR 20.11.3