Fix job_submit/lua support for 'gres' which is now stored as a 'tres'
When requesting jobs so needs a 'gres' prefix.

After 6300d47c2d2 job_desc_msg_t.gres is stored as tres_per_node and
needs a 'gres' prefix.  If the lua script is messing around with that
we need to make sure we add it.

This changes job_submit/lua to keep backward compatibility.

Bug 12642

REVERT 21.08.3