Prevent leaving an empty external launcher step on the step list
If the external launcher's step layout cannot be created, we will bailout.
At this point we need to destroy the partially created step record. If we
don't do this, as we return directly from step_create, we will end up with
a bad step in the step list which can cause a segfault later on.

Ticket 19469.

REVERT with 23.11.7